+++ src/dependencies/dependency.rs
@@ -11,6 +11,7 @@ use std::fmt::{Display, Formatter};
 
 pub struct Dependency {
     pub name: String,
+    pub rename: Option<String>,
     pub version: String,
 
     pub workspace: bool,