+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 * make `*` as a version be a wildcard for `any` - always find prerelease versions
+* handle dependency renames via `package = ""`
 
 ## 0.7.0